use form to refine prompt
